+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. This change
 - rely on kindly/*options* and kindly-advice for options (PR #148)
 - bugfix in handling `:kindly/options` at item preparation
 - check for horizontal in clay and kindly options (PR #149)
+- bugfix in handling options in notebook generation
 
 ## [2-beta15] - 2024-07-27
 - introducing code-and-value and horizontal layout - WIP (PR #127)